Move all physics component logic to the react component file
